feat : 사용자 기본 정보 수정 구현

+++ server/src/main/java/com/albamung/user/service/UserService.java
@@ -39,6 +39,7 @@ public class UserService {
      */
     public String signup(User user) {
         user.setRoles("ROLE_USER");
+        //랜덤 프로필 이미지 저장
         Random random = new Random();
         int num = random.nextInt(100);
         List<String> profileImage = List.of("https://www.gravatar.com/avatar/39f6e2dc52425b1e08027c01bb880be0?s=256&d=identicon&r=PG",
@@ -47,61 +48,35 @@ public class UserService {
                 "https://www.gravatar.com/avatar/e514b017977ebf742a418cac697d8996?s=256&d=identicon&r=PG",
                 "https://www.gravatar.com/avatar/ad240ed5cc406759f0fd72591dc8ca47?s=256&d=identicon&r=PG");
         user.setProfileImage(profileImage.get(num%5));
+
         try {
             userRepository.save(user);
-            return "Welcome";
+            return user.getNickName();
         } catch (Exception e) {
             throw new CustomException("Email or Nick Name is already in use", HttpStatus.UNPROCESSABLE_ENTITY);
         }
     }
 
     /**
-     * 로그인
-     * 엑세스 토큰 및 리프레시 토큰 발급 & 저장
-     */
-//    public TokenDto login(String email, String password) {
-//        try {
-//            //유저 검증
-//            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(email, password));
-//            User user = userRepository.findByEmail(email).orElseThrow();
-//            //refreshToken발급 및 저장
-//            user.setRefreshToken(jwtTokenProvider.createRefreshToken());
-//            //accessToken발급
-//            String access = jwtTokenProvider.createToken(user.getId(), user.getEmail(), user.getRoleList());
-//            //access & refresh 리턴
-//            return new TokenDto(access, user.getRefreshToken());
-//        } catch (AuthenticationException e) {
-//            throw new CustomException("Invalid username/password supplied", HttpStatus.UNPROCESSABLE_ENTITY);
-//        }
-//    }
-
-
-    /**
      * 유저 상세 조회
      */
     @Transactional(readOnly = true)
     public User getUserInfo(Long userId) {
-        User user = verifyUser(userId);
-        if (user.getRoles().equals("ROLE_GUEST")) throw new CustomException("해당 ID는 게스트 유저 입니다", HttpStatus.LOCKED);
-        else return user;
+        return verifyUser(userId);
     }
 
     /**
      * 유저 정보 수정
      */
-    public String putUserInfo(User user, long loginUserId) {
-        User targetUser = verifyUser(user.getId());
-        if (user.getId() != loginUserId)
-            throw new CustomException("You are not the owner of this user", HttpStatus.FORBIDDEN);
-
-        Optional.ofNullable(user.getProfileImage()).ifPresent(targetUser::setProfileImage);
+    public User putUserDefault(User user, long loginUserId) {
+        User targetUser = verifyUser(loginUserId);
         Optional.ofNullable(user.getNickName()).ifPresent(targetUser::setNickName);
-        Optional.ofNullable(user.getAboutMe()).ifPresent(targetUser::setAboutMe);
         Optional.ofNullable(user.getFullName()).ifPresent(targetUser::setFullName);
-        Optional.ofNullable(user.getLocation()).ifPresent(targetUser::setLocation);
-
-        return uri + "users/" + targetUser.getId().toString();
+        Optional.ofNullable(user.getPhone()).ifPresent(targetUser::setPhone);
+        return targetUser;
     }
+
+
 
     /**
      * 유저 존재 검사
@@ -125,19 +100,11 @@ public class UserService {
 //        return new TokenDto(jwtTokenProvider.createToken(user.getId(), user.getEmail(), user.getRoleList()), user.getRefreshToken());
 //    }
 
-    /**
-     * 전체 유저 목록 조회
-     */
-//    @Transactional(readOnly = true)
-//    public Page<User> getUserList(int page) {
-//        PageRequest pageRequest = PageRequest.of(page, 10, Sort.by("id").descending());
-//        return userRepository.findAllByIsGuestIsFalse(pageRequest);
-//    }
 
     public void deleteUser(Long userId, Long loginUserId) {
         User user = verifyUser(userId);
-//        if (user.getId() != loginUserId)
-//            throw new CustomException("You are not the owner of this user", HttpStatus.FORBIDDEN);
+        if (!user.getId().equals(loginUserId))
+            throw new CustomException("You are not the owner of this user", HttpStatus.FORBIDDEN);
         userRepository.delete(user);
     }
 }
